masync: make vdm and future header only
